Update lat and long to double

+++ spring-project/src/main/java/com/example/springproject/CoordinatesDTO.java
@@ -1,23 +1,23 @@
 package com.example.springproject;
 
 public class CoordinatesDTO {
-    private int latitude;
-    private int longitude;
+    private double latitude;
+    private double longitude;
 
     // Getters and setters
-    public int getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(int latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public int getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(int longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 }
